Cant delete meals or items without permission

+++ client/src/components/MealComponent.vue
@@ -25,7 +25,7 @@
             </button>
             <div class="dropdown-menu">
               <button
-                @click="deleteFoodItem(mealData._id, foodItem._id)"
+                @click="deleteFoodItem(mealData._id, foodItem._id, foodItem.profileId._id, mealData.authors)"
                 class="dropdown-item"
               >Delete</button>
             </div>
@@ -78,12 +78,17 @@ export default {
         });
       }
     },
-    deleteFoodItem(mealId, foodItemId) {
-      this.$store.dispatch("removeFoodItem", {
-        mealId,
-        foodItemId,
-        tripId: this.$route.params.tripId
-      });
+    deleteFoodItem(mealId, foodItemId, foodItemProfileId, mealAuthors) {
+      if (
+        this.$store.state.profile._id == foodItemProfileId ||
+        mealAuthors.includes(this.$store.state.user._id)
+      ) {
+        this.$store.dispatch("removeFoodItem", {
+          mealId,
+          foodItemId,
+          tripId: this.$route.params.tripId
+        });
+      } else NotificationService.errorMessage("You can't remove this item");
     }
   },
   computed: {
